Updated
Login button in the header. Upon clicking, it opens a box with a form for both login and register, with a button to switch between them. There is an "X" button that closes this. The register submit button adds data to localStorage and the login button checks if the data matches a user's data